backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.19.0
geopy==2.2.0
Flask-SQLAlchemy==3.0.0